requirements.txt
ibm-watson==3.0.4
python-dotenv==0.10.3
starlette==0.12.1
uvicorn==0.8.2
